refactor(chat): rename router.py to persona_filter.py
Disambiguates from nodes/chat_node:route_chat_node (graph fan-out).
This module's MessageRouter filters which persona_ids should reply to
a message via @mention resolution; the fan-out node handles graph-level
dispatch. Two distinct responsibilities, two distinct names.

Audit conclusion (Task 11):
- agent_stream.py (244) + stream.py (81) — kept separate (combined would
  exceed 300-line single-file limit; orchestrator vs token state machine
  are distinct concerns).
- router.py (58) — renamed to persona_filter.py for clarity.

Spec: docs/superpowers/specs/2026-05-07-dataflow-phase-6-cleanup-design.md §3.1
